Make stdin read only if '-' is provided
Running this in a GitHub check doesn't have a tty which was
causing some UX oddness.

Do the thing I was trying to avoid and just check for '-' as a
special filename to mean stdin.

Also check for `-i` if stdin is provided and fail sicne this
does not make sense.

fixes #17